PHP program to edit docx/doc file
This file edit some contents of an existing docx or doc file.

The basic idea behind this editing is based on the docx/doc file structure. We can say its a archive file including a number of xml and other type of files. We use no docx handling libraries here. We cannot edit docx file directly opening it from the php program. This is because its not a plain file. So we extract the archive file. Then we edit he word/document.xml file which have the contents exactly in the document. 
Just try and enjoy..!